Refactor model types; add Invoice domain models
Refactored several domain model properties from int to byte for fields representing enums or small-range values (e.g., RoomType, StatusId, ScannerType). Added MedicationName to PatientMedicationHistory and VaccineName to PatientVaccination. Expanded mock PatientHealthHistory data in PatientRepository. Introduced new Invoice and InvoiceDetailLine domain models. Removed redundant folder entry from project file.

+++ VKmfSoft_EHealth_API/Models/Domain/Patient/PatientAllergy.cs
@@ -13,7 +13,7 @@ namespace VKmfSoft_EHealth_API.Models.Domain.Patient
         
         [Required(ErrorMessage = "Severity is required.")]
        
-        public required int  SeverityId { get; set; } 
+        public required byte  SeverityId { get; set; } 
        
         public string? Reaction { get; set; }
        
@@ -21,7 +21,7 @@ namespace VKmfSoft_EHealth_API.Models.Domain.Patient
        
         public string? Notes { get; set; }
        
-        public int StatusId { get; set; }
+        public byte StatusId { get; set; }
 
     }
 }
